@@ -1,19 +1,22 @@
 name: qr_mobile_vision
-description: "Plugin for reading QR codes using Google's Mobile Vision API."
-version: 0.3.0
-author: Morgan McKenzie <rmtmckenzie@gmail.com>
+description: "Plugin for reading QR codes using Firebase's Mobile Vision API."
+version: 3.0.0
 homepage: https://github.com/rmtmckenzie/flutter_qr_mobile_vision
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=0.10.15 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.10.0"
 
 flutter:
   plugin:
-    androidPackage: com.github.rmtmckenzie.qrmobilevision
-    pluginClass: QrMobileVisionPlugin
+    platforms:
+      android:
+        package: com.github.rmtmckenzie.qrmobilevision
+        pluginClass: QrMobileVisionPlugin
+      ios:
+        pluginClass: QrMobileVisionPlugin
 
 dependencies:
   flutter:
     sdk: flutter
-  native_device_orientation: ^0.3.0
+  native_device_orientation: ^1.0.0
