@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -7,53 +6,52 @@ import 'package:flutter/rendering.dart';
 import 'package:native_device_orientation/native_device_orientation.dart';
 import 'package:qr_mobile_vision/qr_mobile_vision.dart';
 
-final WidgetBuilder _defaultNotStartedBuilder = (context) => new Text("Camera Loading ...");
-final WidgetBuilder _defaultOffscreenBuilder = (context) => new Text("Camera Paused.");
-final ErrorCallback _defaultOnError = (BuildContext context, Object error) {
+final WidgetBuilder _defaultNotStartedBuilder = (context) => Text("Camera Loading ...");
+final WidgetBuilder _defaultOffscreenBuilder = (context) => Text("Camera Paused.");
+final ErrorCallback _defaultOnError = (BuildContext context, Object? error) {
   print("Error reading from camera: $error");
-  return new Text("Error reading from camera...");
+  return Text("Error reading from camera...");
 };
 
-typedef Widget ErrorCallback(BuildContext context, Object error);
+typedef Widget ErrorCallback(BuildContext context, Object? error);
 
 class QrCamera extends StatefulWidget {
   QrCamera({
-    Key key,
-    @required this.qrCodeCallback,
+    Key? key,
+    required this.qrCodeCallback,
     this.child,
     this.fit = BoxFit.cover,
-    WidgetBuilder notStartedBuilder,
-    WidgetBuilder offscreenBuilder,
-    ErrorCallback onError,
+    WidgetBuilder? notStartedBuilder,
+    WidgetBuilder? offscreenBuilder,
+    ErrorCallback? onError,
     this.formats,
   })  : notStartedBuilder = notStartedBuilder ?? _defaultNotStartedBuilder,
         offscreenBuilder = offscreenBuilder ?? notStartedBuilder ?? _defaultOffscreenBuilder,
         onError = onError ?? _defaultOnError,
-        assert(fit != null),
         super(key: key);
 
   final BoxFit fit;
-  final ValueChanged<String> qrCodeCallback;
-  final Widget child;
+  final ValueChanged<String?> qrCodeCallback;
+  final Widget? child;
   final WidgetBuilder notStartedBuilder;
   final WidgetBuilder offscreenBuilder;
   final ErrorCallback onError;
-  final List<BarcodeFormats> formats;
+  final List<BarcodeFormats>? formats;
 
   @override
-  QrCameraState createState() => new QrCameraState();
+  QrCameraState createState() => QrCameraState();
 }
 
 class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
   }
 
   @override
   dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 
@@ -73,7 +71,7 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
   }
 
   bool onScreen = true;
-  Future<PreviewDetails> _asyncInitOnce;
+  Future<PreviewDetails>? _asyncInitOnce;
 
   Future<PreviewDetails> _asyncInit(num width, num height) async {
     var previewDetails = await QrMobileVision.start(
@@ -112,14 +110,14 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
 
   @override
   Widget build(BuildContext context) {
-    return new LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
+    return LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
       if (_asyncInitOnce == null && onScreen) {
         _asyncInitOnce = _asyncInit(constraints.maxWidth, constraints.maxHeight);
       } else if (!onScreen) {
         return widget.offscreenBuilder(context);
       }
 
-      return new FutureBuilder(
+      return FutureBuilder(
         future: _asyncInitOnce,
         builder: (BuildContext context, AsyncSnapshot<PreviewDetails> details) {
           switch (details.connectionState) {
@@ -131,11 +129,11 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
                 debugPrint(details.error.toString());
                 return widget.onError(context, details.error);
               }
-              Widget preview = new SizedBox(
+              Widget preview = SizedBox(
                 width: constraints.maxWidth,
                 height: constraints.maxHeight,
                 child: Preview(
-                  previewDetails: details.data,
+                  previewDetails: details.data!,
                   targetWidth: constraints.maxWidth,
                   targetHeight: constraints.maxHeight,
                   fit: widget.fit,
@@ -143,17 +141,17 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
               );
 
               if (widget.child != null) {
-                return new Stack(
+                return Stack(
                   children: [
                     preview,
-                    widget.child,
+                    widget.child!,
                   ],
                 );
               }
               return preview;
 
             default:
-              throw new AssertionError("${details.connectionState} not supported.");
+              throw AssertionError("${details.connectionState} not supported.");
           }
         },
       );
@@ -164,63 +162,60 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
 class Preview extends StatelessWidget {
   final double width, height;
   final double targetWidth, targetHeight;
-  final int textureId;
-  final int orientation;
+  final int? textureId;
+  final int? sensorOrientation;
   final BoxFit fit;
 
   Preview({
-    @required PreviewDetails previewDetails,
-    @required this.targetWidth,
-    @required this.targetHeight,
-    @required this.fit,
-  })  : assert(previewDetails != null),
-        textureId = previewDetails.textureId,
-        width = previewDetails.width.toDouble(),
-        height = previewDetails.height.toDouble(),
-        orientation = previewDetails.orientation;
+    required PreviewDetails previewDetails,
+    required this.targetWidth,
+    required this.targetHeight,
+    required this.fit,
+  })   : textureId = previewDetails.textureId,
+        width = previewDetails.width!.toDouble(),
+        height = previewDetails.height!.toDouble(),
+        sensorOrientation = previewDetails.sensorOrientation as int?;
 
   @override
   Widget build(BuildContext context) {
-    double frameHeight, frameWidth;
-
-    return new NativeDeviceOrientationReader(
+    return NativeDeviceOrientationReader(
       builder: (context) {
         var nativeOrientation = NativeDeviceOrientationReader.orientation(context);
 
-        int baseOrientation = 0;
-        if (orientation != 0 && (width > height)) {
-          baseOrientation = orientation ~/ 90;
-          frameWidth = width;
-          frameHeight = height;
-        } else {
-          frameHeight = width;
-          frameWidth = height;
-        }
-
-        int nativeOrientationInt;
+        int nativeRotation = 0;
         switch (nativeOrientation) {
-          case NativeDeviceOrientation.landscapeLeft:
-            nativeOrientationInt = Platform.isAndroid ? 3 : 1;
+          case NativeDeviceOrientation.portraitUp:
+            nativeRotation = 0;
             break;
           case NativeDeviceOrientation.landscapeRight:
-            nativeOrientationInt = Platform.isAndroid ? 1 : 3;
+            nativeRotation = 90;
             break;
           case NativeDeviceOrientation.portraitDown:
-            nativeOrientationInt = 2;
+            nativeRotation = 180;
             break;
-          case NativeDeviceOrientation.portraitUp:
+          case NativeDeviceOrientation.landscapeLeft:
+            nativeRotation = 270;
+            break;
           case NativeDeviceOrientation.unknown:
-            nativeOrientationInt = 0;
+          default:
+            break;
         }
 
-        return new FittedBox(
-          fit: fit,
-          child: new RotatedBox(
-            quarterTurns: baseOrientation + nativeOrientationInt,
-            child: new SizedBox(
-              width: frameWidth,
-              height: frameHeight,
-              child: new Texture(textureId: textureId),
+        int rotationCompensation = ((nativeRotation - sensorOrientation! + 450) % 360) ~/ 90;
+
+        double frameHeight = width;
+        double frameWidth = height;
+
+        return ClipRect(
+          child: FittedBox(
+            fit: fit,
+            child: RotatedBox(
+              quarterTurns: rotationCompensation,
+              child: SizedBox(
+                width: frameWidth,
+                height: frameHeight,
+                child: Texture(textureId: textureId!),
+              ),
             ),
           ),
         );
