@@ -1,15 +1,14 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 class PreviewDetails {
-  num width;
-  num height;
-  num orientation;
-  int textureId;
+  num? width;
+  num? height;
+  num? sensorOrientation;
+  int? textureId;
 
-  PreviewDetails(this.width, this.height, this.orientation, this.textureId);
+  PreviewDetails(this.width, this.height, this.sensorOrientation, this.textureId);
 }
 
 enum BarcodeFormats {
@@ -35,14 +34,14 @@ const _defaultBarcodeFormats = const [
 
 class QrMobileVision {
   static const MethodChannel _channel = const MethodChannel('com.github.rmtmckenzie/qr_mobile_vision');
-  static QrChannelReader channelReader = new QrChannelReader(_channel);
+  static QrChannelReader channelReader = QrChannelReader(_channel);
 
   //Set target size before starting
   static Future<PreviewDetails> start({
-    @required int width,
-    @required int height,
-    @required QRCodeHandler qrCodeHandler,
-    List<BarcodeFormats> formats = _defaultBarcodeFormats,
+    required int width,
+    required int height,
+    required QRCodeHandler qrCodeHandler,
+    List<BarcodeFormats>? formats = _defaultBarcodeFormats,
   }) async {
     final _formats = formats ?? _defaultBarcodeFormats;
     assert(_formats.length > 0);
@@ -50,18 +49,22 @@ class QrMobileVision {
     List<String> formatStrings = _formats.map((format) => format.toString().split('.')[1]).toList(growable: false);
 
     channelReader.setQrCodeHandler(qrCodeHandler);
-    var details = await _channel.invokeMethod(
-        'start', {'targetWidth': width, 'targetHeight': height, 'heartbeatTimeout': 0, 'formats': formatStrings});
+    var details = await _channel.invokeMethod('start', {
+      'targetWidth': width,
+      'targetHeight': height,
+      'heartbeatTimeout': 0,
+      'formats': formatStrings,
+    });
 
     // invokeMethod returns Map<dynamic,...> in dart 2.0
     assert(details is Map<dynamic, dynamic>);
 
-    int textureId = details["textureId"];
-    num orientation = details["surfaceOrientation"];
-    num surfaceHeight = details["surfaceHeight"];
-    num surfaceWidth = details["surfaceWidth"];
+    int? textureId = details["textureId"];
+    num? orientation = details["surfaceOrientation"];
+    num? surfaceHeight = details["surfaceHeight"];
+    num? surfaceWidth = details["surfaceWidth"];
 
-    return new PreviewDetails(surfaceWidth, surfaceHeight, orientation, textureId);
+    return PreviewDetails(surfaceWidth, surfaceHeight, orientation, textureId);
   }
 
   static Future stop() {
@@ -73,14 +76,14 @@ class QrMobileVision {
     return _channel.invokeMethod('heartbeat').catchError(print);
   }
 
-  static Future<List<List<int>>> getSupportedSizes() {
-    return _channel.invokeMethod('getSupportedSizes').catchError(print);
+  static Future<List<List<int>>?> getSupportedSizes() {
+    return _channel.invokeMethod('getSupportedSizes').catchError(print) as Future<List<List<int>>?>;
   }
 }
 
 enum FrameRotation { none, ninetyCC, oneeighty, twoseventyCC }
 
-typedef void QRCodeHandler(String qr);
+typedef void QRCodeHandler(String? qr);
 
 class QrChannelReader {
   QrChannelReader(this.channel) {
@@ -89,7 +92,7 @@ class QrChannelReader {
         case 'qrRead':
           if (qrCodeHandler != null) {
             assert(call.arguments is String);
-            qrCodeHandler(call.arguments);
+            qrCodeHandler!(call.arguments);
           }
           break;
         default:
@@ -99,10 +102,10 @@ class QrChannelReader {
     });
   }
 
-  void setQrCodeHandler(QRCodeHandler qrch) {
+  void setQrCodeHandler(QRCodeHandler? qrch) {
     this.qrCodeHandler = qrch;
   }
 
   MethodChannel channel;
-  QRCodeHandler qrCodeHandler;
+  QRCodeHandler? qrCodeHandler;
 }
