@@ -19,12 +19,16 @@ import android.util.Log;
 import android.util.Size;
 import android.util.SparseIntArray;
 import android.view.Surface;
+import android.view.WindowManager;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+
+import com.google.mlkit.vision.common.InputImage;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-
-import androidx.annotation.NonNull;
 
 import static android.hardware.camera2.CameraMetadata.CONTROL_AF_MODE_AUTO;
 import static android.hardware.camera2.CameraMetadata.CONTROL_AF_MODE_CONTINUOUS_PICTURE;
@@ -35,6 +39,7 @@ import static android.hardware.camera2.CameraMetadata.LENS_FACING_BACK;
  * Implements QrCamera using Camera2 API
  */
 @TargetApi(21)
+@RequiresApi(21)
 class QrCameraC2 implements QrCamera {
 
     private static final String TAG = "cgr.qrmv.QrCameraC2";
@@ -55,13 +60,14 @@ class QrCameraC2 implements QrCamera {
     private ImageReader reader;
     private CaptureRequest.Builder previewBuilder;
     private CameraCaptureSession previewSession;
-    private Size jpegSizes[] = null;
-    private QrDetector2 detector;
-    private int orientation;
+    private Size[] jpegSizes = null;
+    private QrDetector detector;
+    private int sensorOrientation;
     private CameraDevice cameraDevice;
     private CameraCharacteristics cameraCharacteristics;
+    private Frame latestFrame;
 
-    QrCameraC2(int width, int height, Context context, SurfaceTexture texture, QrDetector2 detector) {
+    QrCameraC2(int width, int height, SurfaceTexture texture, Context context, QrDetector detector) {
         this.targetWidth = width;
         this.targetHeight = height;
         this.context = context;
@@ -81,7 +87,36 @@ class QrCameraC2 implements QrCamera {
 
     @Override
     public int getOrientation() {
-        return orientation;
+        // ignore sensor orientation of devices with 'reverse landscape' orientation of sensor
+        // as camera2 api seems to already rotate the output.
+        return sensorOrientation == 270 ? 90 : sensorOrientation;
+    }
+
+    private int getFirebaseOrientation() {
+        WindowManager windowManager = (WindowManager) context.getSystemService(Context.WINDOW_SERVICE);
+        int deviceRotation = windowManager.getDefaultDisplay().getRotation();
+        int rotationCompensation = (ORIENTATIONS.get(deviceRotation) + sensorOrientation + 270) % 360;
+
+        // Return the corresponding FirebaseVisionImageMetadata rotation value.
+        int result;
+        switch (rotationCompensation) {
+            case 0:
+                result = 0;
+                break;
+            case 90:
+                result = 90;
+                break;
+            case 180:
+                result = 180;
+                break;
+            case 270:
+                result = 270;
+                break;
+            default:
+                result = 0;
+                Log.e(TAG, "Bad rotation value: " + rotationCompensation);
+        }
+        return result;
     }
 
     @Override
@@ -115,9 +150,8 @@ class QrCameraC2 implements QrCamera {
         try {
             cameraCharacteristics = manager.getCameraCharacteristics(cameraId);
             StreamConfigurationMap map = cameraCharacteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
-            // it seems as though the orientation is already corrected, so setting to 0
-            // orientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
-            orientation = 0;
+            Integer sensorOrientationInteger = cameraCharacteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
+            sensorOrientation = sensorOrientationInteger == null ? 0 : sensorOrientationInteger;
 
             size = getAppropriateSize(map.getOutputSizes(SurfaceTexture.class));
             jpegSizes = map.getOutputSizes(ImageFormat.JPEG);
@@ -156,15 +190,36 @@ class QrCameraC2 implements QrCamera {
             modes.add(afMode);
         }
 
-        if (modes.contains(CONTROL_AF_MODE_CONTINUOUS_PICTURE)) {
-            return CONTROL_AF_MODE_CONTINUOUS_PICTURE;
-        } else if (modes.contains(CONTROL_AF_MODE_CONTINUOUS_VIDEO)) {
+        if (modes.contains(CONTROL_AF_MODE_CONTINUOUS_VIDEO)) {
             return CONTROL_AF_MODE_CONTINUOUS_VIDEO;
+        } else if (modes.contains(CONTROL_AF_MODE_CONTINUOUS_PICTURE)) {
+            return CONTROL_AF_MODE_CONTINUOUS_PICTURE;
         } else if (modes.contains(CONTROL_AF_MODE_AUTO)) {
             return CONTROL_AF_MODE_AUTO;
         } else {
             return null;
         }
+    }
+
+    static class Frame implements QrDetector.Frame {
+        final Image image;
+        final int firebaseOrientation;
+
+        Frame(Image image, int firebaseOrientation) {
+            this.image = image;
+            this.firebaseOrientation = firebaseOrientation;
+        }
+
+        @Override
+        public InputImage toImage() {
+            return InputImage.fromMediaImage(image, firebaseOrientation);
+        }
+
+        @Override
+        public void close() {
+            image.close();
+        }
+
     }
 
     private void startCamera() {
@@ -180,9 +235,11 @@ class QrCameraC2 implements QrCamera {
         ImageReader.OnImageAvailableListener imageAvailableListener = new ImageReader.OnImageAvailableListener() {
             @Override
             public void onImageAvailable(ImageReader reader) {
-                try (Image image = reader.acquireLatestImage()) {
+                try {
+                    Image image = reader.acquireLatestImage();
                     if (image == null) return;
-                    detector.detect(image);
+                    latestFrame = new Frame(image, getFirebaseOrientation());
+                    detector.detect(latestFrame);
                 } catch (Throwable t) {
                     t.printStackTrace();
                 }
@@ -211,7 +268,6 @@ class QrCameraC2 implements QrCamera {
                     previewBuilder.set(CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_CANCEL);
                 }
             }
-
         } catch (java.lang.Exception e) {
             e.printStackTrace();
             return;
@@ -232,7 +288,6 @@ class QrCameraC2 implements QrCamera {
             }, null);
         } catch (Throwable t) {
             t.printStackTrace();
-
         }
     }
 
@@ -247,7 +302,6 @@ class QrCameraC2 implements QrCamera {
         if (cameraDevice == null) return;
 
         try {
-
             previewSession.setRepeatingRequest(previewBuilder.build(), listener, null);
         } catch (java.lang.Exception e) {
             e.printStackTrace();
@@ -260,6 +314,8 @@ class QrCameraC2 implements QrCamera {
             cameraDevice.close();
         }
         if (reader != null) {
+            if (latestFrame != null) latestFrame.close();
+            latestFrame = null;
             reader.close();
         }
     }
@@ -275,7 +331,7 @@ class QrCameraC2 implements QrCamera {
 
         if (s1.getWidth() > s.getWidth() || s1.getHeight() > s.getHeight()) {
             // ascending
-            if (orientation % 180 == 0) {
+            if (sensorOrientation % 180 == 0) {
                 for (Size size : sizes) {
                     s = size;
                     if (size.getHeight() > targetHeight && size.getWidth() > targetWidth) {
@@ -292,7 +348,7 @@ class QrCameraC2 implements QrCamera {
             }
         } else {
             // descending
-            if (orientation % 180 == 0) {
+            if (sensorOrientation % 180 == 0) {
                 for (Size size : sizes) {
                     if (size.getHeight() < targetHeight || size.getWidth() < targetWidth) {
                         break;
@@ -310,6 +366,4 @@ class QrCameraC2 implements QrCamera {
         }
         return s;
     }
-
-
 }
